Removed the parser mixin
I have removed the code for an expeeriment i made, which made it possible for users to define their own parser in a groovy script field, interpreted at runtime. It works...but i don't think anyone will use it. It accidently got mixed into the 2.0 release train. So this commit should revert that.

It was a proposed solution for people who needed to calculate certain values, which could not be read directly from the map file.