Fix integral division result cast as double
I fixed this by casting the integral division result to double.
This was a findBugs issue that raises a good point because it's a stink in case you care about accuracy of your results.
In our case, we really don't care  because the values are only used to plot out the bars of our graph.
Fixing it the good way would be a serious undertaking and I don't feel like writing a perfectly accurate HexMaths library just to plot out some graphs on a Jenkins build.